Fix warning

+++ apeiron/engine/color_converter.h
@@ -86,9 +86,9 @@ constexpr glm::vec4 as_rgba_norm(std::string_view hex_code)
 constexpr glm::vec3 as_rgb_norm(std::uint8_t r, std::uint8_t g, std::uint8_t b)
 {
   return {
-      r / 255.0f,
-      g / 255.0f,
-      b / 255.0f
+      static_cast<float>(r) / 255.0f,
+      static_cast<float>(g) / 255.0f,
+      static_cast<float>(b) / 255.0f
   };
 }
 
@@ -96,10 +96,10 @@ constexpr glm::vec3 as_rgb_norm(std::uint8_t r, std::uint8_t g, std::uint8_t b)
 constexpr glm::vec4 as_rgba_norm(std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a)
 {
   return {
-      r / 255.0f,
-      g / 255.0f,
-      b / 255.0f,
-      a / 255.0f
+      static_cast<float>(r) / 255.0f,
+      static_cast<float>(g) / 255.0f,
+      static_cast<float>(b) / 255.0f,
+      static_cast<float>(a) / 255.0f
   };
 }
